+++ NEWS.md
@@ -1,5 +1,7 @@
-# ggcall v0.2.9000
+# ggcall v0.3.0.9000
 
+* Added support for ggplot related patchwork operators.
+* Updated `ggcall` attribute of `ggcall()` so it is a language object not a list of such.
 * Improved `ggcall_add_assignments` input validation for a call argument, it has to be a symbol.
 * Edited README file.